% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.qordstan.R
\name{.z_to_cat}
\alias{.z_to_cat}
\title{Transform underlyng latent variable in categorie}
\usage{
.z_to_cat(z, gamma)
}
\arguments{
\item{z}{latent variable}

\item{gamma}{cut points}
}
\value{
categories
}
\description{
Transform underlyng latent variable in categorie
}
\examples{
z = rnorm(100)
gm = rexp(100, 1/2)
z_to_cat(z, gm)
}
