% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_qordstan.R
\name{new_qordstan}
\alias{new_qordstan}
\title{Creates qordstan object}
\usage{
new_qordstan(stan_fit, x, y, q, beta_scale, delta_scale)
}
\arguments{
\item{stan_fit}{stan model fit object}

\item{x}{numeric matrix with covariates}

\item{y}{vector with categories}

\item{q}{quantile}

\item{beta_scale}{standard deviation for the coefs prior}

\item{delta_scale}{standard deviation for the deltas prior (cutpoints)}
}
\value{

}
\description{
Creates qordstan object
}
\examples{

}
