% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.qordstan.R
\name{posterior_resume}
\alias{posterior_resume}
\title{Return vector with average, standard deviation and HDI}
\usage{
posterior_resume(x, cred_mass = 0.95)
}
\arguments{
\item{x}{numeric vector}

\item{cred_mass}{length of creditive interval. Must be between 0 and 1}
}
\value{
numeric vector with mean, standard deviation and HDI
}
\description{
Return vector with average, standard deviation and HDI
}
