% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.qordstan.R
\name{print.qordstan}
\alias{print.qordstan}
\title{Print summary for qordstan object}
\usage{
\method{print}{qordstan}(x, ...)
}
\arguments{
\item{x}{qordstan object}

\item{...}{aditional arguments}
}
\description{
Print summary for qordstan object
}
\examples{
#data = gen_data_example()
#fit = qord_fit(y ~ ., q = 0.5, data = data$example_df, iter = 10, warmup = 5)
#fit
}
