% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.qordstan.R
\name{print.summary.qordstan}
\alias{print.summary.qordstan}
\title{Print method for summary method}
\usage{
\method{print}{summary.qordstan}(x, ...)
}
\arguments{
\item{x}{summary.qordstan object}

\item{...}{additional parameters}
}
\description{
Print method for summary method
}
\examples{
#data = gen_data_example()
#fit = qord_fit(y ~ ., q = 0.5, data = data$example_df, iter = 10, warmup = 5)
#summary(fit)

}
