% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qord_fit.R
\name{qord_fit}
\alias{qord_fit}
\title{Fit the ordinal quantilic regression model using stan}
\usage{
qord_fit(x, y, p, beta_scale = 1, delta_scale = 0.25, ...)
}
\arguments{
\item{x}{numeric matrix with covariates}

\item{y}{vector with categories}

\item{p}{fixed quantile}

\item{beta_scale}{standard deviation for the coefs prior}

\item{delta_scale}{standard deviation for the deltas prior (cutpoints)}
}
\value{
stan model
}
\description{
Fit the ordinal quantilic regression model using stan
}
\examples{
data = gen_data_example()
fit = qord_fit(data$x, data$y, p = 0.5)
}
