% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qordstan-package.R
\docType{package}
\name{qordstan-package}
\alias{qordstan-package}
\alias{qordstan}
\title{Fit bayesian ordinal quantilic regression with Stan.}
\description{
bayesian ordinal quantilic regression with Stan
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
