% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.qordstan.R
\name{summary.qordstan}
\alias{summary.qordstan}
\title{Summary method for qordstan objects}
\usage{
\method{summary}{qordstan}(object, cred_mass = 0.95, ...)
}
\arguments{
\item{object}{a qorstan object}

\item{cred_mass}{length of credibility interval, must be between 0 and 1}

\item{...}{additional arguments}
}
\value{
summary.qordstan object with components:\tabular{ll}{
\code{summary_table} \tab matrix with posterior posterior mean, standard deviation and HDI \cr
\tab \cr
\code{beta_mean} \tab posterior mean of model coefficients \cr
\tab \cr
\code{waic} \tab loo waic object  \cr
\code{n} \tab number of observations \cr
\code{k} \tab number of categories \cr
}
}
\description{
Summary method to qordstan models containing posterior mean, standard deviation
and lower and upper bond of the high density credible interval and other useful quantities
}
\examples{
#data = gen_data_example()
#fit = qord_fit(y ~ ., q = 0.5, data = data$example_df, iter = 10, warmup = 5)
#summary(fit)

}
\seealso{
Other qordstan: 
\code{\link{predict.qordstan}()},
\code{\link{print.summary.qordstan}()},
\code{\link{qord_fit}()}
}
\author{
Pedro Araujo
}
\concept{qordstan}
