% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.qordstan.R
\name{.z_to_cat}
\alias{.z_to_cat}
\title{Transform underlying latent variable in category}
\usage{
.z_to_cat(z, gamma)
}
\arguments{
\item{z}{latent variable}

\item{gamma}{cut points}
}
\value{
vector with numeric categories
}
\description{
Transform underlying latent variable in category
}
