% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qord_fit.R
\name{gen_data_example}
\alias{gen_data_example}
\title{Generate random data to test function}
\usage{
gen_data_example()
}
\value{
list containing X, y, b (real coefs), delta's and gamma's
}
\description{
Generate random data to test function
}
