% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data_example.R
\name{gen_data_example}
\alias{gen_data_example}
\title{Generate random data to test function}
\usage{
gen_data_example()
}
\value{
list containing x (covariates), y (response), b (real coefs),
delta's and gamma's
}
\description{
Generate random data to test function
}
\examples{
data = gen_data_example()

}
