% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_qordstan.R
\name{new_qordstan}
\alias{new_qordstan}
\title{Creates qordstan object}
\usage{
new_qordstan(stan_fit, formula, x, y, q, beta_scale, delta_scale)
}
\arguments{
\item{stan_fit}{stan model fit object}

\item{formula}{model formula}

\item{x}{numeric matrix with covariates}

\item{y}{vector with categories}

\item{q}{quantile}

\item{beta_scale}{standard deviation for the coefs prior}

\item{delta_scale}{standard deviation for the deltas prior (cut-points)}
}
\value{
qordstan object
}
\description{
Creates qordstan object
}
\examples{
#new_qordstan = function(stan_fit, formula, x, y, q, beta_scale, delta_scale)



}
\author{
Pedro Araújo
}
