% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.qordstan.R
\name{predict.qordstan}
\alias{predict.qordstan}
\title{Get posterior predictive sample from the model}
\usage{
\method{predict}{qordstan}(object, type = "cat", new_data = NULL, ...)
}
\arguments{
\item{object}{a qordstan object}

\item{type}{the type of prediction. It will be 'cat' by default, returning
the category. If 'z' returns the underlying latent variable}

\item{new_data}{new covariates for prediction. if null, predicts from model data}

\item{...}{additional parameters}
}
\description{
Generate sample from posterior predictive distribution.
}
\examples{
#data = gen_data_example()
#fit = qord_fit(data$x, data$y, q = 0.5, iter = 10, warmup = 5)
#predict(fit)

}
