% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.qordstan.R
\name{predict.qordstan}
\alias{predict.qordstan}
\title{Get posterior predictive sample from the model}
\usage{
\method{predict}{qordstan}(object, ...)
}
\arguments{
\item{object}{qordstan object}

\item{...}{aditional parameters}

\item{type}{type of prediction. It will be 'cat' by default, returning
the categorie. If 'z' returns the underlyng latent variable}

\item{new_data}{new covariates for prediction.
if null, predicts from model data}
}
\value{

}
\description{
Get posterior predictive sample from the model
}
\examples{
#' data = gen_data_example()
fit = qord_fit(data$x, data$y, q = 0.5, iter = 200, warmup = 100)
predict(fit, type='cat')
}
