% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.qordstan.R
\name{print.qordstan}
\alias{print.qordstan}
\title{Print summary for qord object}
\usage{
\method{print}{qordstan}(x, ...)
}
\arguments{
\item{x}{qordstan object}

\item{...}{aditional arguments}
}
\value{

}
\description{
Print summary for qord object
}
