% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qord_fit.R
\name{qord_fit}
\alias{qord_fit}
\title{Fit the ordinal quantile regression model using Stan}
\usage{
qord_fit(
  x,
  y,
  q,
  beta_scale = 1,
  delta_scale = 0.25,
  iter = 2000,
  warmup = 1000,
  thin = 1,
  chains = 1,
  ...
)
}
\arguments{
\item{x}{numeric matrix with covariates}

\item{y}{vector with categories}

\item{q}{fixed quantile}

\item{beta_scale}{standard deviation for the coefs prior}

\item{delta_scale}{standard deviation for the delta priors (cutpoints)}

\item{iter}{number of iterations}

\item{warmup}{length of warmup sample}

\item{thin}{period of saving samples}

\item{chains}{number of mcmc chains}

\item{...}{aditional rstan parameters}
}
\value{
qordstan object
}
\description{
Fit the ordinal quantile regression model using Stan
}
\examples{
data = gen_data_example()
fit = qord_fit(data$x, data$y, q = 0.5, iter = 10, warmup = 5)
}
