% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.qordstan.R
\name{summary.qordstan}
\alias{summary.qordstan}
\title{Summary method to qordstan models containing posterior mean, standard deviation
and lower and upper bond for creditiva interval}
\usage{
\method{summary}{qordstan}(x, ..., cred_mass)
}
\arguments{
\item{x}{qorstan object}

\item{...}{aditional arguments}

\item{cred_mass}{length of creditive interval, must be between 0 and 1}
}
\value{

}
\description{
Summary method to qordstan models containing posterior mean, standard deviation
and lower and upper bond for creditiva interval
}
\examples{
data = gen_data_example()
qord_model = qord_fit(data$x, data$y, q = 0.5, iter = 10, warmup = 5)
summary(qord_model)

}
