% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.qordstan.R
\name{summary.qordstan}
\alias{summary.qordstan}
\title{Summary method for qordstan objects}
\usage{
\method{summary}{qordstan}(object, cred_mass = 0.95, ...)
}
\arguments{
\item{object}{a qorstan object}

\item{cred_mass}{length of credibility interval, must be between 0 and 1}

\item{...}{aditional arguments}
}
\description{
Summary method to qordstan models containing posterior mean, standard deviation
and lower and upper bond of the high density credible interval
}
\examples{
#data = gen_data_example()
#fit = qord_fit(y ~ ., q = 0.5, data = data$example_df, iter = 10, warmup = 5)
#summary(fit)

}
